Add files via upload
Python notebook for a Random Walk problem.
The problem is as follows:
If you are in the Empire State Building, climbing stairs with dice and playing a game. Rolling the dice, if you get <2, you climb down one step, if you get 3,4,5 you climb up one more. And for 6, you add the amount equals to another random dice throw. Then what is the probability that you would end up at step 60?